+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 [Full Changelog](https://github.com/parse-community/Parse-Swift/compare/2.0.0...2.0.1)
 
 __Fixes__
-- ParseUser should only encode email when User.current?.email is different from current user email  ([#256](https://github.com/parse-community/Parse-Swift/pull/256)), thanks to [Corey Baker](https://github.com/cbaker6).
+- ParseUser should only encode email when User.current?.email is different from current user email ([#256](https://github.com/parse-community/Parse-Swift/pull/256)), thanks to [Corey Baker](https://github.com/cbaker6).
 
 ### 2.0.0
 [Full Changelog](https://github.com/parse-community/Parse-Swift/compare/1.11.0...2.0.0)